Refactor templates to project root
