+++ controlpanel/core/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from .views import IndexView
+
+urlpatterns = [
+    path("", IndexView.as_view(), name="index"),
+]
